Do not run tests for Python 3.5
The test cases for Python 3.5 fail. The reason is that we pass
a `pathlib.Path` object as `output_dir` to Cookiecutter. Internally,
Cookiecutter uses `os.path.join` on this variable, which only accepts
a path-like object since Python 3.6. We could change this in the test
cases, but pathlib is the way forward, and Python 3.5 is only supported
for three more months (end of security fixes: 2020-09-13).