Improve creation and usage of slugified names
Cookiecutter 1.7 includes a slugify extension. Using this is cleaner and
more robust than some string modifications.

Project names and module names do not follow the same naming
conventions. For project names, dashes are preferred while underscores
are mandatory for the module name. This change adds a 'module_slug'
field to accommodate this. The project slug is now also used as the
package name.