Write __version__ to package on creation
Querying the package version from the metadata might provide a better
"single source of truth", but it places a runtime dependency on
setuptools (when using `pkg_resources`). Python 3.8 provides
`importlib.metadata.version` in the standard library, but earlier
versions do not have this.

Writing the version to package during package creation ensures that the
version is the same. The primary downside is that `setuptools_scm`
requires you to specify the path (including module name) in the
`write_to` path, making it less flexible if the name should ever change.