Single source version using importlib.metadata (#13)
In order to reduce the number of places where the version number is defined, do
not let `setuptools_scm` write the `__version__` but instead use
`importlib.metadata` to retrieve the version from the installed package. This
reduces the complexity of the build step and the likelihood of mismatches
between `__version__` and the actual version of the package.

The only downside of this approach is that the module `import.metadata` was
added in Python 3.8. For earlier version, the package `importlib_metadata` can
be used, and this is specified as a conditional requirement in `setup.cfg`.