Move test for __version__ to unit test
Since the `__version__` attribute is no longer written to `_version.py`
during package creation, and is instead retrieved using
`importlib.metadata`, checking to see of the metadata of the package
matches the value of `__version__` is no longer relevant. Instead, the
primary test is now if `__version__` is set properly to some non-default
value.